requirements.txt
Django>=5.1,<5.3
gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
psycopg[binary]>=3.2,<3.3
dj-database-url>=2.3,<3.0
python-dotenv>=1.0,<2.0
openai>=1.54,<2.0
google-api-python-client>=2.151,<3.0
google-auth>=2.36,<3.0
google-auth-httplib2>=0.2.0,<1.0
google-auth-oauthlib>=1.2,<2.0
pypdf>=5.1,<6.0
PyMuPDF>=1.24,<2.0
pytesseract>=0.3,<0.4
Pillow>=11.0,<12.0
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
